fix: clear the 30 warnings -D warnings exposed, incl. a real bug
The tightened clippy gate failed on its first run with 30 pre-existing
warnings. One was a genuine bug.

series_float.rs did 'let _ = tx.send(result)' on an async_channel::Sender from
a std::thread worker. send() there returns a future; nothing polled it, so the
result was discarded and the receiver only woke when tx dropped -- meaning
every SUCCESSFUL series fetch surfaced as 'Fetch worker ended unexpectedly'.
Now send_blocking, matching what metadata_editor.rs already did. It was the
only async_channel send in the tree; the other 'let _ = tx.send' sites are
relm4 senders where send() is synchronous.

Also mine: HomeOut variants carried an Open prefix (clippy::enum_variant_names)
-- renamed to Book/BookDialog/AllBooks, the convention LibraryOut documents --
and thumbs.rs had an unused test helper.

Pre-existing: epub_write.rs kept 7 real functions after its test module (moved
tests to EOF), plus needless borrows, redundant closures, contains-over-any,
sort_by_key, an unnecessary to_string, redundant i32 casts, two redundant
re-bindings, and a complex type now aliased as pages::SelfRebuild (3 sites).

Three suppressed with a written justification rather than contorted: the Lesk
stemmer's identical-bodied suffix rules, build_reader_settings_panel's 8
arguments, and AuthorPageMsg::Fetched -- that last one boxed instead, since
~376 bytes on every message including Refresh is a real cost, not a lint quirk.

+++ src/pages/settings.rs
@@ -609,7 +609,7 @@ fn build_theme_picker(host: &gtk::Grid, catalog: &Arc<Catalog>) {
             continue;
         }
         let block = theme_family_block(&family, name, desc, prefix, active, host, catalog);
-        host.attach(&block, (slot % 2) as i32, (slot / 2) as i32, 1, 1);
+        host.attach(&block, slot % 2, slot / 2, 1, 1);
         slot += 1;
     }
 }
